Updated the dependencies for ITK 5.2
ITK 5.2rc2 is only available via pip currently.
Update to itk >= 5.2 when it is officially released.
Installing itk-io and itk-filtering independently is no longer
supported (?), check later. The install is also much faster, so maybe
there is no need to be extra specific with itk's sub-packages.
Added Python 3.9 to the classifiers (all dependencies agree on the
python versions).